+++ mlflow/cli.py
@@ -2,14 +2,12 @@ import json
 import os
 import sys
 import logging
-import warnings
 
 import click
 from click import UsageError
 
 import mlflow.db
 import mlflow.experiments
-import mlflow.models.cli
 import mlflow.deployments.cli
 import mlflow.projects as projects
 import mlflow.runs
@@ -452,7 +450,6 @@ def gc(backend_store_uri, run_ids):
         print("Run with ID %s has been permanently deleted." % str(run_id))
 
 
-cli.add_command(mlflow.models.cli.commands)
 cli.add_command(mlflow.deployments.cli.commands)
 cli.add_command(mlflow.experiments.commands)
 cli.add_command(mlflow.store.artifact.cli.commands)
@@ -464,16 +461,17 @@ try:
     import pandas  # noqa: E402
     import numpy  # noqa: E402
 except ImportError as e:
-    warnings.warn(
-        "Built in deployment plugins could not be loaded due to missing numpy "
-        "or pandas dependencies. %s." % e
-    )
+    # We are conditional loading these commands since the skinny client does
+    # not support them due to the pandas and numpy dependencies of MLflow Models
+    pass
 else:
+    import mlflow.models.cli
     import mlflow.azureml.cli
     import mlflow.sagemaker.cli
 
     cli.add_command(mlflow.azureml.cli.commands)
     cli.add_command(mlflow.sagemaker.cli.commands)
+    cli.add_command(mlflow.models.cli.commands)
 
 if __name__ == "__main__":
     cli()
